refactor(api): move scorecard reset acceptance in evaluation context
